feat: add follow schema, swagger, refactor dto
