+++ package.json
@@ -8,8 +8,8 @@
   "scripts": {
     "build": "nest build",
     "format": "prettier --write \"src/**/*.ts\" \"test/**/*.ts\"",
-    "start": "nest start -b swc",
-    "start:dev": "nest start --watch -b swc",
+    "start": "nest start -b swc --type-check",
+    "start:dev": "nest start --watch -b swc --type-check",
     "start:debug": "nest start --debug --watch",
     "start:prod": "node dist/main",
     "lint": "eslint \"{src,apps,libs,test}/**/*.ts\" --fix",
@@ -23,11 +23,13 @@
     "@nestjs/common": "^10.0.0",
     "@nestjs/core": "^10.0.0",
     "@nestjs/platform-express": "^10.0.0",
+    "@nestjs/swagger": "^11.0.7",
     "@prisma/client": "^6.4.1",
     "@qdrant/js-client-rest": "^1.13.0",
     "@xenova/transformers": "^2.17.2",
     "aws-sdk": "^2.1692.0",
     "axios": "^1.8.4",
+    "builder-pattern": "^2.2.0",
     "class-transformer": "^0.5.1",
     "class-validator": "^0.14.1",
     "nanoid": "^5.1.3",
