+++ src/main.ts
@@ -1,6 +1,8 @@
 import { NestFactory } from '@nestjs/core';
 import { AppModule } from './app.module';
 import { ValidationPipe } from '@nestjs/common';
+import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
+import metadata from './metadata';
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule);
@@ -12,6 +14,16 @@ async function bootstrap() {
       },
     }),
   );
+
+  const config = new DocumentBuilder()
+    .setTitle('Art Sharing')
+    .setDescription('The Art Sharing API description')
+    .setVersion('1.0')
+    .addTag('artsharing')
+    .build();
+  const documentFactory = () => SwaggerModule.createDocument(app, config);
+  await SwaggerModule.loadPluginMetadata(metadata);
+  SwaggerModule.setup('api', app, documentFactory);
 
   await app.listen(process.env.PORT ?? 3000);
 }
