+++ src/metadata.ts
@@ -1,0 +1,12 @@
+/* eslint-disable */
+export default async () => {
+    const t = {
+        ["./posts/dto/request/media.dto"]: await import("./posts/dto/request/media.dto"),
+        ["./posts/dto/response/media.dto"]: await import("./posts/dto/response/media.dto"),
+        ["./posts/dto/response/user.dto"]: await import("./posts/dto/response/user.dto"),
+        ["./posts/dto/response/category.dto"]: await import("./posts/dto/response/category.dto"),
+        ["./posts/dto/response/post-details.dto"]: await import("./posts/dto/response/post-details.dto"),
+        ["./storage/dto/response.dto"]: await import("./storage/dto/response.dto")
+    };
+    return { "@nestjs/swagger": { "models": [[import("./storage/dto/request.dto"), { "GetPresignedUrlRequestDto": { fileName: { required: true, type: () => String, default: "" }, extension: { required: true, type: () => String, pattern: "/^(png|jpg|jpeg|gif|webp|mp4|mov|avi)$/i" }, mediaType: { required: true, type: () => Object }, directory: { required: false, type: () => String, default: "uncategorized" } } }], [import("./storage/dto/response.dto"), { "GetPresignedUrlResponseDto": { url: { required: true, type: () => String }, key: { required: true, type: () => String } } }], [import("./posts/dto/request/media.dto"), { "MediaDto": { url: { required: true, type: () => String }, media_type: { required: true, type: () => Object } } }], [import("./posts/dto/request/create-post.dto"), { "CreatePostDto": { title: { required: true, type: () => String }, description: { required: false, type: () => String }, medias_data: { required: true, type: () => [t["./posts/dto/request/media.dto"].MediaDto] }, cate_ids: { required: true, type: () => [Number] } } }], [import("./posts/dto/response/media.dto"), { "MediaResponseDto": { id: { required: true, type: () => Number }, media_type: { required: true, type: () => Object }, description: { required: false, type: () => String }, url: { required: true, type: () => String }, creator_id: { required: true, type: () => Number }, downloads: { required: true, type: () => Number }, created_at: { required: true, type: () => Date } } }], [import("./posts/dto/response/user.dto"), { "UserResponseDto": { id: { required: true, type: () => Number }, username: { required: true, type: () => String }, email: { required: true, type: () => String }, password_hash: { required: true, type: () => String }, full_name: { required: true, type: () => String }, profile_picture_url: { required: true, type: () => String }, bio: { required: true, type: () => String }, created_at: { required: true, type: () => Date }, updated_at: { required: true, type: () => Date }, refresh_token: { required: true, type: () => String } } }], [import("./posts/dto/response/category.dto"), { "CategoryResponseDto": { id: { required: true, type: () => Number }, name: { required: true, type: () => String }, cate_name: { required: true, type: () => String }, url: { required: true, type: () => String }, created_at: { required: true, type: () => Date } } }], [import("./posts/dto/response/post-details.dto"), { "PostDetailsResponseDto": { id: { required: true, type: () => Number }, user_id: { required: true, type: () => Number }, title: { required: true, type: () => String }, description: { required: false, type: () => String }, thumbnail_url: { required: true, type: () => String }, is_published: { required: true, type: () => Boolean }, is_private: { required: true, type: () => Boolean }, like_count: { required: true, type: () => Number }, share_count: { required: true, type: () => Number }, comment_count: { required: true, type: () => Number }, created_at: { required: true, type: () => Date }, medias: { required: true, type: () => [t["./posts/dto/response/media.dto"].MediaResponseDto] }, user: { required: true, type: () => [t["./posts/dto/response/user.dto"].UserResponseDto] }, categories: { required: true, type: () => [t["./posts/dto/response/category.dto"].CategoryResponseDto] } } }], [import("./posts/dto/request/update-post.dto"), { "UpdatePostDto": { title: { required: false, type: () => String }, description: { required: false, type: () => String }, cate_ids: { required: false, type: () => [Number] }, medias_data: { required: false, type: () => [t["./posts/dto/request/media.dto"].MediaDto] } } }], [import("./likes/dto/create-like.dto"), { "CreateLikeDto": { target_id: { required: true, type: () => Number }, target_type: { required: true, type: () => Object } } }], [import("./likes/dto/remove-like.dto"), { "RemoveLikeDto": { target_id: { required: true, type: () => Number }, target_type: { required: true, type: () => Object } } }], [import("./shares/dto/create-share.dto"), { "CreateShareDto": { target_id: { required: true, type: () => Number }, target_type: { required: true, type: () => Object }, share_platform: { required: true, type: () => Object } } }], [import("./shares/dto/remove-share.dto"), { "RemoveShareDto": { target_id: { required: true, type: () => Number }, target_type: { required: true, type: () => Object } } }]], "controllers": [[import("./app.controller"), { "AppController": { "getHello": { type: String } } }], [import("./posts/posts.controller"), { "PostsController": { "createPost": { type: t["./posts/dto/response/post-details.dto"].PostDetailsResponseDto }, "updatePost": { type: t["./posts/dto/response/post-details.dto"].PostDetailsResponseDto }, "deletePost": {}, "searchPosts": {}, "getPostDetails": { type: t["./posts/dto/response/post-details.dto"].PostDetailsResponseDto }, "getForYouPosts": {}, "getFollowingPosts": {} } }], [import("./likes/likes.controller"), { "LikesController": { "createLike": {}, "removeLike": {} } }], [import("./shares/shares.controller"), { "SharesController": { "createShare": {}, "removeShare": {} } }], [import("./storage/storage.controller"), { "StorageController": { "getPresignedUrl": { type: t["./storage/dto/response.dto"].GetPresignedUrlResponseDto } } }]] } };
+};
