+++ src/posts/dto/request/update-post.dto.ts
@@ -1,0 +1,24 @@
+import { Type } from "class-transformer";
+import { IsArray, IsInt, IsOptional, IsString, ValidateNested } from "class-validator";
+import { MediaDto } from "./media.dto";
+
+export class UpdatePostDto {
+  @IsOptional()
+  @IsString()
+  title?: string;
+
+  @IsOptional()
+  @IsString()
+  description?: string;
+
+  @IsOptional()
+  @IsArray()
+  @IsInt({ each: true })
+  cate_ids?: number[];
+
+  @IsOptional()
+  @IsArray()
+  @ValidateNested({ each: true })
+  @Type(() => MediaDto)
+  medias_data?: MediaDto[];
+}
