+++ src/posts/dto/response/category.dto.ts
@@ -1,0 +1,7 @@
+export class CategoryResponseDto {
+  id: number;
+  name: string;
+  cate_name: string;
+  url: string;
+  created_at: Date;
+}
